+++ lib/Halcyon/Core.hs
@@ -1,0 +1,12 @@
+{-# LANGUAGE OverloadedStrings #-}
+module Halcyon.Core
+  ( -- * IR Generation
+    module Halcyon.Core.TackyGen
+    -- * Compiler Infrastructure
+  , module Halcyon.Core.Monad
+  , module Halcyon.Core.Settings
+  ) where
+
+import Halcyon.Core.Monad
+import Halcyon.Core.Settings  
+import Halcyon.Core.TackyGen